fix: folders restructuring
