+++ src/elasticai/experiment_framework/cli.py
@@ -1,4 +1,5 @@
 import click
+
 import elasticai.experiment_framework.synthesis as synth
