--- src/elasticai/experiment_framework/remote_control/basic_usage/__init__.py
@@ -1,3 +1,0 @@
-from .main import main
-
-__all__ = ["main"]
