+++ src/elasticai/experiment_framework/remote_control/connection_provider.py
@@ -1,11 +1,11 @@
 import asyncio
 import logging
 
-from ..protocol.constants import (
+from .constants import (
     NUM_MAX_RETRIES,
     RETRY_DELAY_SECONDS,
 )
-from ..protocol.io_stream import IOStream
+from .io_stream import IOStream
 from .tcp_protocol_stream import TCPProtocolStream
 
 _logger = logging.getLogger(__name__)
@@ -25,6 +25,7 @@ class ConnectionProvider:
 
         for attempt in range(max_trials):
             try:
+
                 def factory():
                     return TCPProtocolStream(self._on_lost)
 
@@ -38,6 +39,6 @@ class ConnectionProvider:
                 await asyncio.sleep(RETRY_DELAY_SECONDS)
         _logger.error(f"[CLIENT] fail to connect after {max_trials}:")
         raise ConnectionError(f"[CLIENT] fail to connect after {max_trials}")
-    
+
     def _on_lost(self, exc):
         _logger.warning(f"connection lost: {exc}")
