+++ src/elasticai/experiment_framework/remote_control/constants.py
@@ -1,5 +1,5 @@
-from enum import Enum
 import struct
+from enum import Enum
 
 SYNC_BYTE = 0xAA
 HEADER_FORMAT = "<BBBBH"
@@ -7,8 +7,11 @@ HEADER_SIZE = struct.calcsize(HEADER_FORMAT)
 NUM_MAX_RETRIES = 10
 RETRY_DELAY_SECONDS = 1.0
 MAX_TRANSACTIONS = 0xFF
+MAX_PAYLOAD_lEN = 0xFFFF
 
 CONNECTION_TIMEOUT_SECONDS = 10
+
+RESPONSE_TIMEOUT = 5.0
 
 NUM_BYTES_FOR_ID = 1
