+++ src/elasticai/experiment_framework/remote_control/header.py
@@ -1,9 +1,15 @@
-# header.py
 import struct
 from dataclasses import dataclass
+
 from .commands import Command
+from .constants import (
+    HEADER_FORMAT,
+    HEADER_SIZE,
+    MAX_TRANSACTIONS,
+    SYNC_BYTE,
+    MAX_PAYLOAD_lEN,
+)
 from .flags import Flags
-from .constants import SYNC_BYTE, HEADER_FORMAT, HEADER_SIZE, MAX_TRANSACTIONS
 
 
 @dataclass
@@ -13,27 +19,6 @@ class Header:
     transaction_id: int
     payload_len: int
 
-    @property
-    def total_size(self) -> int:
-        """Full message size including payload"""
-        return HEADER_SIZE + self.payload_len
-
-    @property
-    def get_command(self):
-        return self.command
-
-    @property
-    def get_flages(self):
-        return self.flags
-
-    @property
-    def get_transaction_id(self):
-        return self.transaction_id
-
-    @property
-    def get_payload_len(self):
-        return self.payload_len
-
     def __post_init__(self) -> None:
         if not isinstance(self.command, Command):
             raise TypeError(f"expected Command, got {type(self.command)}")
@@ -41,8 +26,8 @@ class Header:
             raise TypeError(f"expected Flags, got {type(self.flags)}")
         if not 0 <= self.transaction_id <= MAX_TRANSACTIONS:
             raise ValueError(f"transaction_id out of range: {self.transaction_id}")
-        if self.payload_len < 0:
-            raise ValueError(f"payload_len cannot be negative: {self.payload_len}")
+        if not 0 <= self.payload_len <= MAX_PAYLOAD_lEN:
+            raise ValueError(f"payload_len out of range: {self.payload_len}")
 
     def to_bytes(self) -> bytes:
         return struct.pack(
@@ -56,7 +41,7 @@ class Header:
 
     @classmethod
     def from_bytes(cls, data: bytes | bytearray | memoryview) -> "Header":
-        if len(data) < HEADER_SIZE:
+        if len(data) != HEADER_SIZE:
             raise ValueError(f"too short: need {HEADER_SIZE} bytes, got {len(data)}")
         if data[0] != SYNC_BYTE:
             raise ValueError(f"invalid sync byte: {data[0]:#x}")
