+++ src/elasticai/experiment_framework/remote_control/helpers.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Any, Dict
+from typing import Any, Dict, Literal
 
 from .commands import Command
 from .constants import NUM_BYTES_FOR_ID
@@ -8,7 +8,10 @@ from .message import Message
 _logger = logging.getLogger(__name__)
 
 
-def interpret_message(msg: Message, byte_order: str = "little") -> Dict[str, Any]:
+def interpret_message(
+    msg: Message, 
+    byte_order: Literal["little", "big"] = "little"
+    ) -> Dict[str, Any]:
     result: Dict[str, Any] = {
         "command": msg.header.command,
         "transaction_id": msg.header.transaction_id,
@@ -41,16 +44,16 @@ def interpret_message(msg: Message, byte_order: str = "little") -> Dict[str, Any
                 result["data"] = {"func_id": func_id, "payload": extra}
             else:
                 result["data"] = None
-                _logger.warning("OPEN_TASK insufficient payload")
+                _logger.warning("[CLIENT]OPEN_TASK insufficient payload")
 
         case Command.DATA_CHUNK:
             if len(msg.payload) >= NUM_BYTES_FOR_ID:
                 data_id = int.from_bytes(msg.payload[:NUM_BYTES_FOR_ID], byte_order)
-                chunk = msg.payload[NUM_BYTES_FOR_ID:]  # ← no task_id
+                chunk = msg.payload[NUM_BYTES_FOR_ID:]
                 result["data"] = {"data_id": data_id, "chunk": chunk}
             else:
                 result["data"] = None
-                _logger.warning("DATA_CHUNK insufficient payload")
+                _logger.warning("[CLIENT]DATA_CHUNK insufficient payload")
 
         case Command.RETURN:
             result["data"] = msg.payload
@@ -60,13 +63,13 @@ def interpret_message(msg: Message, byte_order: str = "little") -> Dict[str, Any
 
         case _:
             result["data"] = msg.payload
-            _logger.warning("unknown command: %s", msg.header.command)
+            _logger.warning("[CLIENT]unknown command: %s", msg.header.command)
 
-    _logger.debug("interpreted: %s", result)
+    _logger.debug("[CLIENT] interpreted: %s", result)
     return result
 
 
-def format_message(msg: Message, byte_order: str = "little") -> str:
+def format_message(msg: Message, byte_order: Literal["little", "big"] = "little") -> str:
     header = msg.header
     parts = [
         f"command={header.command.name}",
@@ -94,7 +97,7 @@ def format_message(msg: Message, byte_order: str = "little") -> str:
                 data_id = int.from_bytes(
                     msg.payload[:NUM_BYTES_FOR_ID], byte_order, signed=False
                 )
-                chunk = msg.payload[NUM_BYTES_FOR_ID:]  # ← no task_id
+                chunk = msg.payload[NUM_BYTES_FOR_ID:]
                 parts += [
                     f"data_id={data_id}",
                     f"chunk={chunk.hex() or '<empty>'}",
