--- src/elasticai/experiment_framework/remote_control/network_layer/__init__.py
@@ -1,4 +1,0 @@
-from .connection import Connection
-from .internal_transport_protocol import InternalTransportProtocolTCP
-
-__all__ = ["Connection", "InternalTransportProtocolTCP"]
