--- src/elasticai/experiment_framework/remote_control/pc_side/__init__.py
@@ -1,7 +1,0 @@
-from .protocol import (
-    TaskManager,
-)
-
-__all__ = [
-    "TaskManager",
-]
