--- src/elasticai/experiment_framework/remote_control/pc_side/network_layer/__init__.py
@@ -1,4 +1,0 @@
-from .connection_provider import ConnectionProvider
-from .tcp_protocol_stream import TCPProtocolStream
-
-__all__ = ["ConnectionProvider", "TCPProtocolStream"]
