--- src/elasticai/experiment_framework/remote_control/pc_side/protocol/__init__.py
@@ -1,20 +1,0 @@
-from .commands import Command
-from .exceptions import *
-from .flags import Flags
-from .header import Header
-from .io_stream import IOStream
-from .message import Message
-from .message_io import MessageIO
-from .task import Task
-from .task_manager import TaskManager
-
-__all__ = [
-    "Command",
-    "Flags",
-    "Message",
-    "Header",
-    "Task",
-    "IOStream",
-    "MessageIO",
-    "TaskManager",
-]
