--- src/elasticai/experiment_framework/remote_control/pc_side/protocol/constants.py
@@ -1,29 +1,0 @@
-import struct
-from enum import Enum
-
-SYNC_BYTE = 0xAA
-HEADER_FORMAT = "<BBBBH"
-HEADER_SIZE = struct.calcsize(HEADER_FORMAT)
-NUM_MAX_RETRIES = 10
-RETRY_DELAY_SECONDS = 1.0
-MAX_TRANSACTIONS = 0xFF
-MAX_PAYLOAD_lEN = 0xFFFF
-
-CONNECTION_TIMEOUT_SECONDS = 10
-
-RESPONSE_TIMEOUT = 5.0
-
-NUM_BYTES_FOR_ID = 1
-
-MAX_MSG_ID = 255
-
-
-class TransportType(Enum):
-    TCP = 0x01
-    UDP = 0x02
-    SERIAL = 0x03
-
-
-MAX_CONNECTED_DEVICES = 4
-
-NUM_BYTES_OFFSET_DATA_ID_IN_PAYLOAD = 0
