+++ src/elasticai/experiment_framework/synthesis/synthesis.py
@@ -1,14 +1,14 @@
-import logging
-import warnings
-from collections.abc import Mapping, Iterator, Iterable
-import os
 import dataclasses
+import logging
+import os
+import warnings
 from abc import abstractmethod
+from collections.abc import Iterable, Iterator, Mapping
 from dataclasses import dataclass
 from enum import StrEnum, auto
-from pathlib import Path
-from typing import Protocol, Self, runtime_checkable, override
 from hashlib import blake2s
+from pathlib import Path
+from typing import Protocol, Self, override, runtime_checkable
 
 
 class TargetPlatforms(StrEnum):
