+++ tests/unit/message_io_test.py
@@ -1,7 +1,5 @@
-
-
-from elasticai.experiment_framework.remote_control.pc_side.protocol.commands import Command
-from elasticai.experiment_framework.remote_control.pc_side.protocol.message import Message
+from elasticai.experiment_framework.remote_control.commands import Command
+from elasticai.experiment_framework.remote_control.message import Message
 
 
 def make_return(payload: bytes) -> bytes:
